+++ test/odd_job/registry_test.exs
@@ -5,7 +5,7 @@ defmodule OddJob.RegistryTest do
 
   describe "via/2" do
     test "returns a name suitable for use in :via registration" do
-      assert via(:job, "pool sup") == {:via, Registry, {OddJob.Registry, {:job, "pool sup"}}}
+      assert via(:job, :pool_sup) == {:via, Registry, {OddJob.Registry, {:job, :pool_sup}}}
     end
   end
 end