+++ test/odd_job/utils_test.exs
@@ -1,4 +1,4 @@
 defmodule OddJob.UtilsTest do
-  use ExUnit.Case
+  use ExUnit.Case, async: true
   doctest OddJob.Utils
 end